requirements.txt
ffmpeg-python
tensorflow==1.15
deeplabcut
streamlit
